[PATCH] Fix a few 32/64 bit warnings

+++ NQ/r_misc.c
@@ -381,8 +381,9 @@ R_SetupFrame(void)
 	if ((surface_p - surfaces) > r_maxsurfsseen)
 	    r_maxsurfsseen = surface_p - surfaces;
 
-	Con_Printf("Used %d of %d surfs; %d max\n", surface_p - surfaces,
-		   surf_max - surfaces, r_maxsurfsseen);
+	Con_Printf("Used %d of %d surfs; %d max\n",
+		   (int)(surface_p - surfaces),
+		   (int)(surf_max - surfaces), r_maxsurfsseen);
     }
 
     if (r_numedges.value) {
